Persistant rooms are updated separately from normal room updates, and regardless of their area being empty.
Tpstat was missing some ! operators when checking for tokens on the target.

Removed some hardcoded immortal names from various commands, requiring future updates whenever proper role designation for different abilities are fully fleshed out.

When a token spell is cast, the level of the caster is no longer set on the 4th value of the token, as that can be checked with the level ifcheck.