Added "force_brief" portal flag.  Causes those entering a portal to act as if BRIEF were set for the look afterward.
Added PREWIMPY/WIMPY triggers.  They operate in the same manner as PREFLEE/FLEE triggers, except are done during a wimpy flee instead of a normal flee.
Added REMORT script command:
  Syntax: remort $PLAYER

  Purpose: prompts the player to select the starting remort class.

Added MULTICLASS and REMORT triggers, called when the player multiclasses and remorts, respectively.

Modified the script trigger table to decouple the type used and the index in the table.  This will help prevent triggers being inoperable due to ordering mistakes in the table.