Added TABARD slot.  It is inherently NORESTRING.  Item is meant to be used as a reward or special function.  The NORESTRING is to keep forgeries from being passed around.
Restring allows color changes to object shorts that are NORESTRING (including tabards).  The name field for such objects are not changed.
Fixed a few memory bugs with restring.
Added a save for the name field for restring.

Added a BITVECTOR entity type with fields referring to valid flags on the bitvector table.  The output value is a BOOLEAN.
Example: $(OBJECT.extra.glow) returns a TRUE if the object is glowing.